refactor(team-mode): split live delivery facade
